Add files via upload
On that day Dr. Muhammad Iqbal Tariq divided the class into groups of 5 to dissemble and then re-assemble a desktop. We were required to take parts out of the desktop one after another: Desktop cover, wires, battery, RAM, Hard-disc, and lastly the CPU fan. Then, we were told to reassemble the desktop again and the desktop had to work when powered on. The first group to finish would be rewarded. Unfortunately, we weren't the fastest but our desktop worked. I found this activity very helpful as I learned the interiors of a desktop.