Add files via upload
I learned from this proposal that Drs. are humans too just like me. I had in mind they were some kind of superior humans that I was afraid to approach to interview them. After I interviewed Dr. Alif Ridzuan, I realized that Drs. are very kind. Dr. Alif was very sweet and he did the interview with us providing all the necessary information that we needed! I also learned how it is working in a group. The proposal was complicated to make, but I can't say that I didn't enjoy making it.